Power Implemented for player

+++ src/main/java/com/brandomine/tech/common/capability/leveling/LevelInfo.java
@@ -3,9 +3,9 @@ package com.brandomine.tech.common.capability.leveling;
 import net.minecraft.nbt.NBTTagCompound;
 
 public class LevelInfo {
-	private int xp = 0;
-	private int level = 1;
-	private int maxXp = 20;
+	private double xp = 0;
+	private double level = 1;
+	private double maxXp = 20;
 	
 	public LevelInfo(LevelStorage level){
 		this.level = this.getLevel();
@@ -13,62 +13,62 @@ public class LevelInfo {
 		this.maxXp = this.getMaxXp();
 	}
 	
-	public LevelInfo(int level, int xp, int maxXp){
+	public LevelInfo(double level, double xp, double maxXp){
 		this.level = level;
 		this.maxXp = maxXp;
 		this.xp = xp;
 	}
 	
-	public LevelInfo(int xp, int maxXp) {
+	public LevelInfo(double xp, double maxXp) {
 		this(1, xp, maxXp);
 	}
 
 
-	public int getLevel(){
+	public double getLevel(){
 		return level;
 	}
 	
-	public int getXp(){
+	public double getXp(){
 		return xp;
 	}
 	
-	public int getMaxXp(){
+	public double getMaxXp(){
 		return maxXp;
 	}
 	
 	public void writeToNBT(NBTTagCompound tag){
 		NBTTagCompound nbt = new NBTTagCompound();
-		nbt.setInteger("Level", this.getLevel());
-		nbt.setInteger("Xp", this.getXp());
-		nbt.setInteger("MaxXp", this.getMaxXp());
+		nbt.setDouble("Level", this.getLevel());
+		nbt.setDouble("Xp", this.getXp());
+		nbt.setDouble("MaxXp", this.getMaxXp());
 		tag.setTag("LevelInfo", nbt);
 	}
 	
-	public void setLevel(int level){
+	public void setLevel(double level){
 		this.level = level;
 	}
 	
-	public void setMaxXp(int maxXp){
+	public void setMaxXp(double maxXp){
 		this.maxXp = maxXp;
 	}
 	
-	public void setXp(int xp){
+	public void setXp(double xp){
 		this.xp = xp;
 	}
 	
 	public void readFromNBT(NBTTagCompound tag){
 		NBTTagCompound nbt = tag.getCompoundTag("LevelInfo");
-		level = nbt.getInteger("Level");
-		xp = nbt.getInteger("Xp");
-		maxXp = nbt.getInteger("MaxXp");
+		level = nbt.getDouble("Level");
+		xp = nbt.getDouble("Xp");
+		maxXp = nbt.getDouble("MaxXp");
 	}
 	
-	public void levelUp(int levels){
+	public void levelUp(double levels){
 		this.level += levels;
 	}
 	
-	public void addXp(int amount){
-		int runAmount = amount;
+	public void addXp(double amount){
+		double runAmount = amount;
 		while(runAmount != 0){
 			this.xp = xp + 1;
 			if(xp == maxXp){
